app.py
from flask import Flask
from flask_script import Manager

from FlaskServer import FlaskServer

app = Flask(__name__)
manager = Manager(app)

manager.add_command('runserver', FlaskServer())


@app.route('/')
def hello_world():
    return 'Hello World!'


if __name__ == '__main__':
    manager.run()
