requirements.txt
gpiozero==1.6.2
RPi.GPIO==0.7.0
requests==2.25.1
importlib==1.0.4
git+https://github.com/sn4k3/FakeRPi
rpi==0.0.1
Flask-Script==2.0.6
python-decouple==3.4